requirements.txt
fastapi
uvicorn
pydantic
python-multipart
httpx
websockets
openai-whisper
numpy
aiofiles
googletrans==4.0.0rc1
soundfile
gtts
openai-whisper 
pydub 
ffmpeg-python
python-docx 
pymupdf 
pytesseract 
pillow
sqlalchemy
cloudinary
openai
python-dotenv
